refactor: altera tipo de valorServico para BigDecimal para garantir precisão financeira

+++ src/main/java/com/barbearia/api/dto/RelatorioComissaoDTO.java
@@ -1,10 +1,11 @@
 package com.barbearia.api.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record RelatorioComissaoDTO(
-        Double totalFaturado,
-        Double totalComissao,
+        BigDecimal totalFaturado,
+        BigDecimal totalComissao,
         Integer quantidadeServicos
 ) {
 }
